refactor: extract markdown pipeline and segmented-control, fix svg time axis font scaling
