+++ src/components/notes/NoteMarkdownPreview.tsx
@@ -23,126 +23,7 @@ import { MermaidDiagram } from "./MermaidDiagram";
 import { CodeBlock } from "./CodeBlock";
 import { Callout } from "./Callout";
 import { MathBlock } from "./MathBlock";
-
-// ── Remark plugins (shared with note-editor) ──────────────────────────────────
-
-const CALLOUT_RE = /^\[!([^\]]+)\]([\+\-]?)([\s\S]*)/;
-
-const remarkCallout = () => (tree: MdastRoot) => {
-  visit(tree, "blockquote", (node: Blockquote) => {
-    const first = node.children[0] as Paragraph | undefined;
-    const firstText = first?.children?.[0] as MdastText | undefined;
-    if (!firstText || firstText.type !== "text") return;
-    const match = CALLOUT_RE.exec(firstText.value);
-    if (!match) return;
-    const [, type, foldMark, rest] = match;
-    const collapsible = foldMark === "+" || foldMark === "-";
-    const defaultOpen = foldMark !== "-";
-    const titleText = rest.trim().split("\n")[0].trim();
-    if (rest.trim()) {
-      firstText.value = rest.trim().replace(/^[^\n]*\n?/, "");
-      if (!firstText.value) first!.children.shift();
-    } else {
-      node.children.shift();
-    }
-    node.data = node.data ?? {};
-    node.data.hName = "callout";
-    node.data.hProperties = {
-      "data-callout-type": type.toLowerCase(),
-      "data-title": titleText || type,
-      "data-collapsible": String(collapsible),
-      "data-default-open": String(defaultOpen),
-    };
-  });
-};
-
-const remarkPromoteDisplayMath = () => (tree: MdastRoot) => {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  visit(tree as any, "paragraph", (node: Paragraph, index: number | undefined, parent: MdastRoot | undefined) => {
-    if (!parent || index === undefined) return;
-    if (node.children.length !== 1) return;
-    const child = node.children[0] as InlineMath | undefined;
-    if (child?.type !== "inlineMath") return;
-    (parent as MdastRoot).children.splice(index, 1, {
-      type: "math",
-      value: child.value,
-      data: { hName: "div", hProperties: { className: ["math", "math-display"] } },
-    } as unknown as MdastRoot["children"][number]);
-  });
-};
-
-function makeLatexPlugins() {
-  const latexBlocks: string[] = [];
-
-  const rehypeCaptureLatex = () => (tree: Root) => {
-    latexBlocks.length = 0;
-    visit(tree, "element", (node: Element) => {
-      const cls = (node.properties?.className as string[] | undefined) ?? [];
-      if (cls.includes("math-display")) {
-        const text = (node.children[0] as Text | undefined)?.value ?? "";
-        latexBlocks.push(text);
-      }
-    });
-  };
-
-  const rehypeMergedPass = () => (tree: Root) => {
-    let i = 0;
-    visit(tree, (node, index: number | undefined, parent: Parent | undefined) => {
-      if (node.type === "element") {
-        const cls = ((node as Element).properties?.className as string[] | undefined) ?? [];
-        if (cls.includes("katex-display")) {
-          if (latexBlocks[i] !== undefined) {
-            (node as Element).tagName = "mathblock";
-            (node as Element).properties = { "data-latex": latexBlocks[i++] };
-          }
-          return SKIP;
-        }
-      }
-      if (
-        node.type === "text" &&
-        (node as Text).value.includes("==") &&
-        parent &&
-        index !== undefined
-      ) {
-        const text = (node as Text).value;
-        const parts = text.split(/(==.+?==)/g);
-        if (parts.length > 1) {
-          const nodes: ElementContent[] = parts
-            .map((part): ElementContent | null => {
-              if (part.startsWith("==") && part.endsWith("==") && part.length > 4) {
-                return { type: "element", tagName: "mark", properties: {}, children: [{ type: "text", value: part.slice(2, -2) }] } as Element;
-              }
-              if (part === "") return null;
-              return { type: "text", value: part } as Text;
-            })
-            .filter((n): n is ElementContent => n !== null);
-          (parent as Parent).children.splice(index, 1, ...nodes);
-          return index;
-        }
-      }
-    });
-  };
-
-  return { rehypeCaptureLatex, rehypeMergedPass };
-}
-
-// ── InlineCode ────────────────────────────────────────────────────────────────
-
-const COLOR_RE = /^(#([0-9a-fA-F]{3,4}|[0-9a-fA-F]{6}|[0-9a-fA-F]{8})|(rgba?|hsla?)\s*\([^)]+\))$/i;
-
-function InlineCode({ className, children }: { className?: string; children?: React.ReactNode }) {
-  if (className?.startsWith("language-")) return <>{children}</>;
-  const text = String(children ?? "").trim();
-  const isColor = COLOR_RE.test(text);
-  return (
-    <code className="px-1 py-0.5 rounded bg-[var(--surface-3)] font-mono text-[0.786rem] text-[var(--text-primary)]">
-      {isColor && (
-        <span style={{ display: "inline-block", width: "0.7em", height: "0.7em", borderRadius: "50%", background: text, border: "1px solid color-mix(in srgb, var(--text-primary) 20%, transparent)", marginRight: "0.35em", verticalAlign: "middle" }} />
-      )}
-      {children}
-    </code>
-  );
-}
+import { remarkCallout, remarkPromoteDisplayMath, makeLatexPlugins, InlineCode } from "@/lib/markdown/pipeline";
 
 // ── NoteMarkdownPreview ───────────────────────────────────────────────────────
